Move imports to new base mixin file
This commit only moves the missing elements to base.py. No code changes
are in the moved or copied lines.
